# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout/session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create a Stripe checkout session for VIP subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionRequest"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /discord/oauth/url:
    get:
      operationId: getDiscordOAuthUrl
      tags: [discord]
      summary: Get Discord OAuth URL for account linking
      responses:
        "200":
          description: OAuth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthUrl"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /discord/status:
    get:
      operationId: getDiscordStatus
      tags: [discord]
      summary: Get Discord connection status for the current user
      responses:
        "200":
          description: Discord connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /discord/disconnect:
    post:
      operationId: disconnectDiscord
      tags: [discord]
      summary: Unlink Discord account
      responses:
        "200":
          description: Discord account unlinked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /discord/sync-logs:
    get:
      operationId: getDiscordSyncLogs
      tags: [discord]
      summary: List recent Discord sync log entries (admin)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: failuresOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Discord sync log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordSyncLogsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    OAuthUrl:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    DiscordSyncLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        discordUserId:
          type: ["string", "null"]
        action:
          type: string
        stripeEvent:
          type: ["string", "null"]
        success:
          type: boolean
        errorMessage:
          type: ["string", "null"]
        attempts:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - action
        - success
        - attempts
        - createdAt
    DiscordSyncLogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/DiscordSyncLog"
        failureCount:
          type: integer
        totalCount:
          type: integer
      required:
        - logs
        - failureCount
        - totalCount
    DiscordStatus:
      type: object
      properties:
        connected:
          type: boolean
        discordUserId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
      required:
        - connected
...
    CheckoutSessionRequest:
      type: object
      properties:
        priceType:
          type: string
          enum: [monthly, yearly, premium, premium-yearly]
        couponCode:
          type: string
      required:
        - priceType
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...